requirements.txt
opencv_python==4.5.3.56
qt_material==2.8.8
numpy==1.19.5
Pillow==8.3.1
PyQt5==5.15.4
